Add vitest and time utils tests
